little change

+++ admin_manageusers.php
@@ -55,8 +55,10 @@ if(!$_SESSION['is_auth']) {
                 userjson = data;
                 var i = 0;
                 $.each(data, function(key,val) {
-                    var btnState = ""
-                    if(val.validate) { btnState = 'disabled="true"' };
+                    var btnState = "";
+		    var btnClass = "";
+                    if(val.validate) { btnState = 'disabled="true"' }
+		    else{ btnClass="btn-primary" };
 
                     $("#tbl_users tbody").append(
                     '<tr data=' + i + '>'+
@@ -64,7 +66,7 @@ if(!$_SESSION['is_auth']) {
                     '<td>' + val.firstName + '</td>'+
                     '<td>'+ val.lastName +'</td>'+
                     '<td><a href="mailto:' + val.email + '">' + val.email + '</a></td>'+
-                    '<td><a href="#"><button class="btn btn-primary validate "' + btnState + 'onClick="validateUser('+i+')">Validate</button></a> ' +
+                    '<td><a href="#"><button class="btn ' + btnClass + ' validate "' + btnState + 'onClick="validateUser('+i+')">Validate</button></a> ' +
                     '<a href="#"><button class="btn btn-danger" onClick="deleteUser('+i+')">Delete</button></a></td>'
                     +'</tr>');
                     i++;
@@ -78,40 +80,47 @@ if(!$_SESSION['is_auth']) {
     
     function deleteUser(id) {
         
-        var userdelete = userjson[id];
-        $.ajax({
-            url: "http://devgrenoble.senslab.info/rest/admin/user",
-            type: "DELETE",
-            contentType: "application/json",
-            data: JSON.stringify(userdelete),
-            dataType: "text",
+	if(confirm("Delete user?"))
+        {
+        	var userdelete = userjson[id];
+        	$.ajax({
+           	url: "http://devgrenoble.senslab.info/rest/admin/user",
+            	type: "DELETE",
+            	contentType: "application/json",
+            	data: JSON.stringify(userdelete),
+            	dataType: "text",
             
-            success:function(data){
-                $("tr[data="+id+"]").remove()    
-        },
-            error:function(XMLHttpRequest, textStatus, errorThrows){
-                alert("error: " + errorThrows)
-            }
-        });
+            	success:function(data){
+                	$("tr[data="+id+"]").remove()    
+        	},
+            	error:function(XMLHttpRequest, textStatus, errorThrows){
+                	alert("error: " + errorThrows)
+            	}
+       		});
+	}
     };
     
     
     function validateUser(id) {
         
-        var uservalidate = userjson[id];
-        $.ajax({
-            url: "http://devgrenoble.senslab.info/rest/admin/user?validate",
-            type: "POST",
-            dataType: "text",
-            contentType: "application/json; charset=utf-8",
-            data: JSON.stringify(uservalidate),
-            success:function(data){
-                 $("tr[data="+id+"] .validate").attr("disabled","disabled")
-        },
-            error:function(XMLHttpRequest, textStatus, errorThrows){
-                alert("error:" + errorThrows)
-            }
-        });
+	if(confirm("Validate user?"))
+	{
+        	var uservalidate = userjson[id];
+        	$.ajax({
+            	url: "http://devgrenoble.senslab.info/rest/admin/user?validate",
+            	type: "POST",
+          	dataType: "text",
+            	contentType: "application/json; charset=utf-8",
+            	data: JSON.stringify(uservalidate),
+            	success:function(data){
+                	$("tr[data="+id+"] .validate").attr("disabled","disabled")
+			$("tr[data="+id+"] .validate").removeClass("btn-primary");
+        	},
+            	error:function(XMLHttpRequest, textStatus, errorThrows){
+                	alert("error:" + errorThrows)
+            	}
+	})
+        };
     };
     
     </script>
